modules/pub_egress/requirements.txt
structlog==21.5.0
environs==6.0.0
python-dateutil==2.8.1
pandas==1.2.5
